requirements.txt
openai>=1.50.0
streamlit>=1.36.0
pandas>=2.2.2
python-dotenv>=1.0.1
google-api-python-client>=2.170.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
requests>=2.32.3
beautifulsoup4>=4.12.3
